Fix crash when reactor is deregistered
Fix crash if a reactor is deregistered in the middle of an update cycle just before its turn to be serviced.  This resets the update cycle index so the iteration is not corrupted. (Not sure if this will repeat the update cycle immediately, or if some reactors won't be updated until the next cycle.)