Add support for non-burner reactors
Remove the check for reactor burner in `reactor_built()`.  (It was already missing in `hotkey_pressed()`.  Add the check for `reactor.burner` in `tick_interfaces()`.  If the reactor has no burner, then the fuel signals will simply be zero forever.  The temperature and enable signals function normally and are very useful to measure stored energy and control the operation.